+++ src/v2/ion/i18nTransformer.js
@@ -46,7 +46,7 @@ import Bundles from 'util/Bundles';
 import Logger from 'util/Logger';
 import { getAuthenticatorDisplayName } from '../view-builder/utils/AuthenticatorUtil';
 import { FORMS, AUTHENTICATOR_KEY } from './RemediationConstants';
-import { I18N_BASE_ATTRIBUTE_ENROLL_PROFILE_MAPPINGS } from './utils/i18nBaseAttributeMappings';
+import { I18N_BASE_ATTRIBUTE_ENROLL_PROFILE_MAPPINGS } from '../view-builder/views/enroll-profile/i18nBaseAttributeMappings';
 
 const WEBAUTHN_API_GENERIC_ERROR_KEY = 'authfactor.webauthn.error';
